Implement basic meta-rl algorithm for NAS over MLPs.
Implement a basic meta-rl based on probabilstic decision trees.
This allows us to sample multple actions from a given policy in a given state.
It allows us to se RL algorithms by giving sane log_probs.

Current issues with our tree-based actor are that it generate huge values.
This is proabbly because theinputs to the actor are huge (un-normalized sizes of NN layers).
We will need to find a better nomailzation scheme, and I expect results to improve.

Moving forward, we will add an actor that generates only CNNs.
Lastly, we can make an actor that generates ether CNNs or MLPs by composing existing solutions.

Most of the `magic` of this aproach relates to computing log_probs in tree.ptree.