Reshape reward function.
Previously, we rewarded our networks linearly for their accuracy.
This is nonsensical, as 50% accuracy is more impressive with 100 labels rather than 2.
So, (1/# labels) ought to be out 0.
As we diverge from (1/n), we should approach infinity.
Extreme (in)accuracies ought to be rewarded accordingly.

See #7 for details on reward reshaping.