Normalize batch values and respect space limits.
A central issue in accuracy so far has been poor batch normalization.
The values going in to the MLP actor are huge (~500).
Large values beget large updates...gradient explosions.
However, our environment needs state values to be  translatable to neural network layers.

The solution is maintain our current internal representation, and tranform it for the neural net.
Additionally, we would like to be able to control the allowed ranges for each part of the state.
That is, the range on chanels ought to be independent from ther dilation.

To this end, we have introduced `fields`.
Fields are any neural network parameter than our sytem needs to generate.
A field consists of a range, a transformation, and a good initial value.

WARNING to future devs:
    RuntimeError: pad should be smaller than half of kernel size, but got padW =
If you got here via grep, I'm sorry.
Our lack of mathematical rigor in nodes.py finally bit you.
I hope this didn't ruin a long running experiment.
Either way, I owe you a beer.

Resolves #1, #5, #6.