Fix `nan` poisoning in MLP networks.
Interval for MLP size did not include `0`.
This was because I didn't want MLP to generate NN's with a layer of size 0.
However, field mapping turned 0 in to an absurdly large number.

This made log probs = -inf.
Attempting VPG with inf log probs causes network to be filled with NANs.

Issue still affects CNN, but this is a start.